+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- core/RenderContext : add setSwapchainParameters()
 - multibody : enable displaying forces applies to frame (with lifetime) (https://github.com/Simple-Robotics/candlewick/pull/93)
 - core/DebugScene : add function to add arrow entities (https://github.com/Simple-Robotics/candlewick/pull/93)
 - core : Add `LambertMaterial` (https://github.com/Simple-Robotics/candlewick/pull/92)